JWT sign and verify extension points
