+++ blueocean-pipeline-api-impl/src/test/java/io/jenkins/blueocean/rest/impl/pipeline/PipelineBaseTest.java
@@ -45,7 +45,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.LogManager;
 
-import static io.jenkins.blueocean.auth.jwt.JwtToken.X_BLUEOCEAN_JWT;
+import static io.jenkins.blueocean.auth.jwt.impl.JwtTokenImpl.X_BLUEOCEAN_JWT;
 import static org.junit.Assert.fail;
 
 /**
